+++ R/methods-coerce.R
@@ -88,12 +88,10 @@ NULL
 #' @export
 as.data.table.SpatVector <- function(x, keep.rownames = FALSE, geom = NULL,
     include_values = TRUE, ...) {
-    # if looking for polygon XY...
-    if (terra::is.polygons(x)) {
-        if (!is.null(geom)) {
-            if (geom == "XY") {
-                return(.spatvector_to_dt(x, include_values = include_values))
-            }
+    if (isTRUE(toupper(geom) == "XY")) {
+        # if looking for polygon XY...
+        if (terra::is.polygons(x)) {
+            return(.spatvector_to_dt(x, include_values = include_values))
         }
     }
     # all other conditions: pass to terra then set as DT
@@ -106,6 +104,19 @@ as.data.table.SpatVector <- function(x, keep.rownames = FALSE, geom = NULL,
 #' @method as.data.table giottoPolygon
 #' @export
 as.data.table.giottoPolygon <- function(x, ...) {
+    # DF conversion with "XY" not supported by {terra} with nrow 0
+    # nrow 0 also loses geom type, so specific code has to be here
+    if (nrow(x) == 0L) {
+        base <- terra::as.data.frame(x[]) |> data.table::setDT()
+        geom_cols <- data.table::data.table(
+            geom = integer(),
+            part = integer(),
+            x = numeric(),
+            y = numeric(),
+            hole = integer()
+        )
+        return(cbind(geom_cols, base))
+    }
     as.data.table(x[], ...)
 }
 
@@ -113,6 +124,16 @@ as.data.table.giottoPolygon <- function(x, ...) {
 #' @method as.data.table giottoPoints
 #' @export
 as.data.table.giottoPoints <- function(x, ...) {
+    # DF conversion with "XY" not supported by {terra} with nrow 0
+    # nrow 0 also loses geom type, so specific code has to be here
+    if (nrow(x) == 0L) {
+        base <- terra::as.data.frame(x[]) |> data.table::setDT()
+        geom_cols <- data.table::data.table(
+            x = numeric(),
+            y = numeric()
+        )
+        return(cbind(base, geom_cols))
+    }
     as.data.table(x[], ...)
 }